feat: Add cron entrypoint
Create a second entrypoint, `/sbin/entrypoint_cron.sh`, that allows
a second instance of container that will run
`/etc/cron.daily/apt-cacher-ng` which gets installed by the package.
Ideally this is done in a `docker-compose.yml` file but can also be
a separate `docker` command.  Both methods are documented in `README.md`

* Dockerfile: Install cron package and add `entrypoint_cron.sh`
  Remove cron jobs not needed.
  When you specify `--entrypoint /sbin/entrypoint_cron.sh` it will
  override what is set in `Dockerfile`.  This ensures full backwards
  compatibility at the cost of the image being slightly larger now that
  `cron` is installed.

* docker-compose.yml: Updated to use cron entrypoint

* entrypoint_cron.sh: New file
  Runs cron in the foreground and print logs to stdout for easy
  inspection with docker.

* README.md: Document use of cron entrypoint
  Includes using just `docker` as well as `docker-compose`